backend/requirements.txt
Flask==1.1.1
psutil==5.6.7
psycopg2==2.8.4
pyavagen==0.3.3
eventlet==0.25.1
gunicorn==20.0.4
flask-cors==3.0.8
flask-wtf==0.14.2
flask-login==0.4.1
flask-admin==1.5.4
flask-restx==0.1.0
flask-moment==0.9.0
flask-migrate==2.5.2
flask-argon2==0.1.5.1
python-dotenv==0.10.5
flask_socketio==4.2.1
flask-sqlalchemy==2.4.1
flask-jwt-extended==3.24.1
